carta1

+++ super_trunfo.c
@@ -7,7 +7,7 @@ char codigo[4]; //A01
 char nomeCidade[50]; //Riodejaneiro
 int numero_habitantes; // 200.000 
 float areaTotal; // 2.000.000 km²
-double pib; // 50.000.000 bilhões
+float pib; // 50.000.000 bilhões
 int pontos_turisticos1; // 50
 
 printf("Digite o estado: \n");
